backend/requirements.txt
Flask==3.1.3
flask_cors==6.0.2
flask_limiter==4.1.1
geopy==2.4.1
mysql_connector_repackaged
numpy
pandas
PyJWT
python-dotenv
scikit_learn
Werkzeug